feat(datepicker): changed month in monthYearLabel to long

+++ src/lib/core/datetime/native-date-formats.ts
@@ -10,7 +10,7 @@ export const GH_NATIVE_DATE_FORMATS: GhDateFormats = {
   },
   display: {
     dateInput: {year: 'numeric', month: 'numeric', day: 'numeric'},
-    monthYearLabel: {year: 'numeric', month: 'short'},
+    monthYearLabel: {year: 'numeric', month: 'long'},
     dateA11yLabel: {year: 'numeric', month: 'long', day: 'numeric'},
     monthYearA11yLabel: {year: 'numeric', month: 'long'},
   }
